+++ angular-workspace/projects/ni/nimble-angular/table-column/duration-text/public-api.ts
@@ -1,0 +1,2 @@
+export * from './nimble-table-column-duration-text.directive';
+export * from './nimble-table-column-duration-text.module';